implementing sending file when view command called
Co-authored by João Costa jmmcostinha@gmail.com
Co-authored by João Fraga joao.biscaia.fraga@gmail.com

+++ src/Tintolmarket.java
@@ -1,9 +1,11 @@
 
 import java.io.BufferedInputStream;
 import java.io.BufferedOutputStream;
+import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
+import java.io.FileReader;
 import java.io.InputStream;
 import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
@@ -57,22 +59,38 @@ public class Tintolmarket {
                 if (splitCommand[0].equals("add") || splitCommand[0].equals("a")) {
 
                     outStream.writeObject(command); //send command
-
-                    File fileToSend = new File(CLIENT_DIR + splitCommand[2]);
-                    FileInputStream fileInputStream = new FileInputStream(fileToSend);
-                    InputStream inputFile = new BufferedInputStream(fileInputStream);
-                
-                    byte[] buffer = new byte[1024];
-                    outStream.writeObject((int) fileToSend.length());
-                    int bytesRead;
                     
-                    // enviar ficheiro   7100bytes
-                    while ((bytesRead = inputFile.read(buffer)) != -1) {
-                        outStream.write(buffer, 0, bytesRead); //send file
-                        outStream.flush();
-                    }
+                    BufferedReader br = new BufferedReader(new FileReader("./data_bases/wines.txt"));
+                    String line;
+                    boolean exists = false;
 
-                    inputFile.close();                    
+                    //check if wine to be added exists in wine.txt
+                    while ((line = br.readLine()) != null && !exists) {
+                        if (line.contains(splitCommand[1])) {
+                            exists = true;
+                        }
+                    }
+                    br.close();
+                    
+                    //only send image file if wine doesn't exist
+                    if(!exists) {
+
+                        File fileToSend = new File(CLIENT_DIR + splitCommand[2]);
+                        FileInputStream fileInputStream = new FileInputStream(fileToSend);
+                        InputStream inputFile = new BufferedInputStream(fileInputStream);
+                    
+                        byte[] buffer = new byte[1024];
+                        outStream.writeObject((int) fileToSend.length());
+                        int bytesRead;
+                        
+                        // enviar ficheiro
+                        while ((bytesRead = inputFile.read(buffer)) != -1) {
+                            outStream.write(buffer, 0, bytesRead); //send file
+                            outStream.flush();
+                        }
+
+                        inputFile.close();      
+                    }              
 
                 } else if (splitCommand[0].equals("view") || splitCommand[0].equals("v")) {
 
@@ -82,32 +100,50 @@ public class Tintolmarket {
                     int bytesRead;
                     byte[] buffer = new byte[1024];
 
-                    File f = new File(CLIENT_DIR + splitCommand[1]);
-                    FileOutputStream fileOutputStream = new FileOutputStream(f);
-                    OutputStream outputFile = new BufferedOutputStream(fileOutputStream);
+                    BufferedReader br = new BufferedReader(new FileReader("./data_bases/wines.txt"));
+                    String line;
+                    String lineForSplit = null;
+                    boolean exists = false;
 
-                    try {
-                        fileSize = (int) inStream.readObject();
-                        int totalsize = fileSize;
-
-                        //receive file
-                        while (totalsize > 0 ) {
-                            if(totalsize >= 1024) {
-                                bytesRead = inStream.read(buffer,0,1024);
-                                System.out.println(bytesRead);
-                            } else {
-                                bytesRead = inStream.read(buffer,0,totalsize);
-                                System.out.println(bytesRead);
-                            }
-                            outputFile.write(buffer,0,bytesRead);
-                            totalsize -= bytesRead;
-                            outputFile.flush();
+                    //check if wine to be shown exists in wine.txt
+                    while ((line = br.readLine()) != null && !exists) {
+                        if (line.contains(splitCommand[1])) {
+                            exists = true;
+                            lineForSplit = line;
                         }
+                    }
+                    br.close();
 
-                        outputFile.close();
-                        fileOutputStream.close();
-                    } catch (ClassNotFoundException e) {
-                        e.printStackTrace();
+                    if (exists) {
+                        String[] splitLine = lineForSplit.split(";");//wineName;wineImage
+
+                        File f = new File(CLIENT_DIR +  splitLine[1]);
+                        FileOutputStream fileOutputStream = new FileOutputStream(f);
+                        OutputStream outputFile = new BufferedOutputStream(fileOutputStream);
+
+                        try {
+                            fileSize = (int) inStream.readObject();
+                            int totalsize = fileSize;
+
+                            //receive file
+                            while (totalsize > 0 ) {
+                                if(totalsize >= 1024) {
+                                    bytesRead = inStream.read(buffer,0,1024);
+                                    System.out.println(bytesRead);
+                                } else {
+                                    bytesRead = inStream.read(buffer,0,totalsize);
+                                    System.out.println(bytesRead);
+                                }
+                                outputFile.write(buffer,0,bytesRead);
+                                totalsize -= bytesRead;
+                                outputFile.flush();
+                            }
+
+                            outputFile.close();
+                            fileOutputStream.close();
+                        } catch (ClassNotFoundException e) {
+                            e.printStackTrace();
+                        }
                     }
 
                 } else {
